Meow Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
pillow
torch
transformers
requests